requirements.txt
requests
mysql-connector-python
python-dotenv
boto3